# docs/swagger.yaml
basePath: /
definitions:
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_online:
        type: boolean
      last_seen:
        type: string
      updated_at:
        type: string
      username:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - email
    - username
    type: object
host: 127.0.0.1
info:
  contact: {}
  description: Boom boom — and into production.
  title: Shat API
  version: 0.001.001
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Аутентифікація користувача та отримання токена
      parameters:
      - description: Дані для входу
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Користувач та токен
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Невірний логін або пароль
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Вхід користувача
      tags:
      - auth
  /api/v1/auth/profile:
    get:
      consumes:
      - application/json
      description: Отримати інформацію про поточного автентифікованого користувача
      produces:
      - application/json
      responses:
        "200":
          description: Профіль користувача
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Неавторизований доступ
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Користувача не знайдено
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - UserTokenAuth: []
      summary: Отримати профіль користувача
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Створення облікового запису користувача
      parameters:
      - description: Дані для реєстрації
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Статус успішної реєстрації
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Невірні дані
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Реєстрація нового користувача
      tags:
      - auth
securityDefinitions:
  UserTokenAuth:
    in: header
    name: X-User-Token
    type: apiKey
swagger: "2.0"